HARMONY-1378: Refactored to move docs generation code into subdir + fixed caching
