+++ app/frontends/docs/docs.ts
@@ -10,10 +10,10 @@ import inc from 'markdown-it-include';
 import mark from 'markdown-it-mark';
 import replaceLink from 'markdown-it-replace-link';
 import toc from 'markdown-it-toc-done-right';
-import HarmonyRequest from '../models/harmony-request';
-import { getRequestRoot } from '../util/url';
-import env from '../util/env';
-import version from '../util/version';
+import HarmonyRequest from '../../models/harmony-request';
+import { getRequestRoot } from '../../util/url';
+import env from '../../util/env';
+import version from '../../util/version';
 import { promisify } from 'util';
 import { generateServicesDocs } from './service-docs-markdown-it-plugin';
 import { interpolate } from './interpolation-markdown-it-plugin';
@@ -29,6 +29,7 @@ const PROD_EDL = 'https://urs.earthdata.nasa.gov';
 const PROD_COLLECTION_ID = 'C1940472420-POCLOUD';
 const UAT_COLLECTION_ID = 'C1234208438-POCLOUD';
 
+// cached generated documentation html
 let docsHtml;
 
 /**
@@ -41,7 +42,7 @@ let docsHtml;
  * counters with.
  * @returns An object with two properties: tableCount and exampleCount.
  */
-async function getTableAndExampleCounts(): Promise<{ tableCount: number, exampleCount: number }> {
+async function getTableAndExampleCounts(): Promise<{ tableCount: number, exampleCount: number; }> {
   const markdownFiles = await readDir(MARKDOWN_DIR);
   let tableCount = 0;
   let exampleCount = 0;
@@ -73,6 +74,92 @@ function markdownInterpolate(token: string, mappings: { [key: string]: () => str
 }
 
 /**
+ * reads the markdown files, parses them, and then renders them to HTML
+ * @param root - The root of the URL for the environment.
+ * @returns a promise that resolves to a string.
+ */
+async function generateDocumentation(root: string): Promise<string> {
+  let { tableCount, exampleCount } = await getTableAndExampleCounts();
+  let edlHost = UAT_EDL;
+  let exampleCollectionId = UAT_COLLECTION_ID;
+  if (root === PROD_ROOT) {
+    edlHost = PROD_EDL;
+    exampleCollectionId = PROD_COLLECTION_ID;
+  }
+  // markdown parser
+  const md = new MarkDownIt(
+    {
+      html: true,
+      linkify: true,
+      highlight: function (str, lang): string {
+        if (lang && hljs.getLanguage(lang)) {
+          try {
+            return '<pre class="hljs"><code>' +
+              hljs.highlight(str, { language: lang, ignoreIllegals: true }).value +
+              '</code></pre>';
+          } catch (__) { }
+        }
+        return '<pre class="hljs"><code>' + md.utils.escapeHtml(str) + '</code></pre>';
+      },
+    },
+  )
+    // interpolate some values into the inline tags
+    .use(inline, 'root_replace', 'text', function (tokens, idx) {
+      tokens[idx].content = markdownInterpolate(tokens[idx].content, {
+        root: () => root,
+        exampleCounter: () => `${exampleCount--}`,
+        tableCounter: () => `${tableCount--}`,
+      });
+    })
+    // interpolate values in non-inline tags
+    .use(interpolate, {
+      edl: () => edlHost.slice(8),
+      exampleCollection: () => exampleCollectionId,
+      root: () => root,
+    })
+    // add 'copy' button to code blocks
+    .use(copy, {
+      btnText: 'COPY',
+      successText: 'COPIED',
+      showCodeLanguage: true,
+    })
+    // replace edl or root links with proper value for the current environment
+    .use(replaceLink, {
+      processHTML: true,
+      replaceLink: function (link: string, _env: string, _token: string, _htmlToken: string) {
+        if (link === 'edl') {
+          return edlHost;
+        }
+        if (link.startsWith('%7B%7Broot%7D%7D')) {
+          return link.replace('%7B%7Broot%7D%7D', root);
+        }
+        return link;
+      },
+    })
+    // Add anchor tags to headers
+    .use(anchor, {
+      permalink: true,
+      permalinkBefore: true,
+      permalinkSymbol: '§',
+    })
+    // Add support for using '==<text>==' to mark up text
+    .use(mark)
+    .use(toc, {
+      listType: 'ul',
+      level: 2,
+    })
+    // Create the documentation for the services
+    .use(generateServicesDocs, {})
+    // Add support for importing markdown fragments into other markdown files
+    .use(inc, {
+      root: 'app/markdown/',
+    });
+
+  const markDown = (await readFile(path.join(MARKDOWN_DIR, 'docs.md'))).toString('utf-8');
+  return md.render(markDown);
+}
+
+/**
  * Express.js handler that returns the Harmony documentation page content.
  *
  * @param req - The request sent by the client
@@ -80,86 +167,8 @@ function markdownInterpolate(token: string, mappings: { [key: string]: () => str
  */
 export default async function docsPage(req: HarmonyRequest, res: Response): Promise<void> {
   const root = getRequestRoot(req);
-  docsHtml = null;
   if (!docsHtml) {
-    let { tableCount, exampleCount } = await getTableAndExampleCounts();
-    let edlHost = UAT_EDL;
-    let exampleCollectionId = UAT_COLLECTION_ID;
-    if (root === PROD_ROOT) {
-      edlHost = PROD_EDL;
-      exampleCollectionId = PROD_COLLECTION_ID;
-    }
-    // markdown parser
-    const md = new MarkDownIt(
-      {
-        html: true,
-        linkify: true,
-        highlight: function (str, lang): string {
-          if (lang && hljs.getLanguage(lang)) {
-            try {
-              return '<pre class="hljs"><code>' +
-                hljs.highlight(str, { language: lang, ignoreIllegals: true }).value +
-                '</code></pre>';
-            } catch (__) { }
-          }
-          return '<pre class="hljs"><code>' + md.utils.escapeHtml(str) + '</code></pre>';
-        },
-      },
-    )
-      // interpolate some values into the inline tags
-      .use(inline, 'root_replace', 'text', function (tokens, idx) {
-        tokens[idx].content = markdownInterpolate(tokens[idx].content, {
-          root: () => root,
-          exampleCounter: () => `${exampleCount--}`,
-          tableCounter: () => `${tableCount--}`,
-        });
-      })
-      // interpolate values in non-inline tags
-      .use(interpolate, {
-        edl: () => edlHost.slice(8),
-        exampleCollection: () => exampleCollectionId,
-        root: () => root,
-      })
-      // add 'copy' button to code blocks
-      .use(copy, {
-        btnText: 'COPY',
-        successText: 'COPIED',
-        showCodeLanguage: true,
-      })
-      // replace edl or root links with proper value for the current environment
-      .use(replaceLink, {
-        processHTML: true,
-        replaceLink: function (link: string, _env: string, _token: string, _htmlToken: string) {
-          if (link === 'edl') {
-            return edlHost;
-          }
-          if (link.startsWith('%7B%7Broot%7D%7D')) {
-            return link.replace('%7B%7Broot%7D%7D', root);
-          }
-          return link;
-        },
-      })
-      // Add anchor tags to headers
-      .use(anchor, {
-        permalink: true,
-        permalinkBefore: true,
-        permalinkSymbol: '§',
-      })
-      // Add support for using '==<text>==' to mark up text
-      .use(mark)
-      .use(toc, {
-        listType: 'ul',
-        level: 2,
-      })
-      // Create the documentation for the services
-      .use(generateServicesDocs, {})
-      // Add support for importing markdown fragments into other markdown files
-      .use(inc, {
-        root: 'app/markdown/',
-      });
-
-    const markDown = (await readFile(path.join(MARKDOWN_DIR, 'docs.md'))).toString('utf-8');
-    docsHtml = md.render(markDown);
+    docsHtml = await generateDocumentation(root);
   }
 
   // render the mustache templates + the rendered markdown
@@ -170,3 +179,5 @@ export default async function docsPage(req: HarmonyRequest, res: Response): Prom
     docsHtml,
   });
 }
+
+
