+++ app/markdown/overview.md
@@ -1,6 +1,6 @@
 ## Overview
 Harmony provides access to services that can transform data from NASA's [Earth Observing Systems Data and Information System (EOSDIS)](https://www.earthdata.nasa.gov/eosdis/) Distributed Active Archive Centers (DAAC).
-Transformations can be requested using one of two [Open Geospatial Consortium (OCG)](https://www.ogc.org/) inspired APIs.
+Transformations can be requested using one of two [Open Geospatial Consortium (OGC)](https://www.ogc.org/) inspired APIs.
 
 Data processed by Harmony is staged in [AWS](https://aws.com) S3 buckets owned by NASA or optionally in user owned S3 buckets. Harmony provides signed URLs or temporary access credentials to users for data staged in NASA buckets.
